% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGMmfpt-package.R
\docType{package}
\name{CGMmfpt-package}
\alias{CGMmfpt}
\alias{CGMmfpt-package}
\title{CGMmfpt: transition and HMM methods for CGM data}
\description{
CGMmfpt estimates participant-level empirical or hidden Markov transition
matrices and summarizes them using mean first-passage time (MFPT). The HMM
uses beta-uniform mixture emission templates with optional overlap around
adjacent state boundaries.
}
\author{
\strong{Maintainer}: Denise M. Scholtens \email{dscholtens@northwestern.edu}

Authors:
\itemize{
  \item Yuanzhi Yu
  \item Margaret M. Banker
}

}
\keyword{internal}
